+++ src/work_functions.cpp
@@ -150,12 +150,12 @@ void SelectEffect(the_lamp_state_t *lamp_state) {
 }
 
 void convertRAW(the_lamp_state_t *lamp_state) {
-    lamp_state->brightness_raw = (lamp_state->brightness+1) >> 4;
-    lamp_state->speed_raw      = ((512-lamp_state->speed) >> 4) +1;
+    lamp_state->brightness_raw = (uint8_t)(lamp_state->brightness+1) >> 4;
+    lamp_state->speed_raw      = (uint8_t)((512-lamp_state->speed) >> 4) +1;
     if (lamp_state->effect <= EFF_CODE_SNAKE )
-        lamp_state->scale_raw  = lamp_state->scale;
+        lamp_state->scale_raw  = (uint8_t)lamp_state->scale;
     else
-        lamp_state->scale_raw  = lamp_state->scale / 10;
+        lamp_state->scale_raw  = (uint8_t)lamp_state->scale / 10;
     DPRINTLN("convertRAW: done");
 }
